requirements.txt
flask
google-genai
datetime